Fastest Algorithm to find Palindromes upto n
This algorithm is the fastest algorithm which is possible to find the all possible prime numbers between 1 to n. This Algorithm is correct you can verify this by commenting printing logic of primes by just printing size of array and using the formula given in this link by putting n in it (https://primes.utm.edu/howmany.html).
This runs fairly quickly on normal pc with in a second for n = 10^8. Further we can use this algo by just using multiple threads to run normal pc at its full extent.